feat: add start popUp
