+++ src/index.js
@@ -2,6 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './sass/index.css';
 import App from './App';
+// import StartPopUp from './StartPopUp';
+
 
 ReactDOM.render(<App />, document.getElementById('root'));
