+++ src/sass/components/_modal.scss
@@ -17,3 +17,13 @@
   background: #fff;
   height: 200px;
 }
+
+// .btnContainer {
+
+//   display: flex;
+//   justify-content: center;
+//   top: 0;
+//   right: 0;
+//   bottom: 0;
+//   left: 0;
+// }
